refactor : Repository 구현체에 따르도록 변경
