+++ src/test/java/com/example/bak/company/domain/CompanyRepositoryStub.java
@@ -1,12 +1,13 @@
 package com.example.bak.company.domain;
 
 
+import com.example.bak.company.application.command.port.CompanyCommandPort;
 import com.example.bak.global.support.AbstractStubRepository;
 import java.util.Objects;
 
 public class CompanyRepositoryStub
         extends AbstractStubRepository<Long, Company>
-        implements CompanyRepository {
+        implements CompanyCommandPort {
 
     @Override
     protected Long getId(Company company) {
